+++ pallets/ddc-staking/src/lib.rs
@@ -15,12 +15,12 @@ use codec::{Decode, Encode, HasCompact};
 use frame_support::{
 	dispatch::Codec,
 	parameter_types,
-	traits::{Currency, DefensiveSaturating, LockIdentifier, WithdrawReasons},
+	traits::{Currency, DefensiveSaturating, LockIdentifier, UnixTime, WithdrawReasons},
 	BoundedVec,
 };
 use scale_info::TypeInfo;
 use sp_runtime::{
-	traits::{AtLeast32BitUnsigned, CheckedSub, Saturating, Zero},
+	traits::{AtLeast32BitUnsigned, Saturating, Zero},
 	RuntimeDebug,
 };
 
@@ -29,6 +29,14 @@ use sp_std::prelude::*;
 
 pub use pallet::*;
 
+/// Two minutes.
+///
+/// If you are changing this, check `on_finalize` hook to ensure `CurrentEra` is capable to hold the
+/// value with the new era duration.
+const DDC_ERA_DURATION_MS: u128 = 120_000;
+
+/// 2023-01-01 00:00:00 UTC
+const DDC_ERA_START_MS: u128 = 1_672_531_200_000;
 const DDC_STAKING_ID: LockIdentifier = *b"ddcstake"; // DDC maintainer's stake
 
 /// The balance type of this pallet.
@@ -102,18 +110,6 @@ impl<AccountId, Balance: HasCompact + Copy + Saturating + AtLeast32BitUnsigned +
 	}
 }
 
-/// Preference of what happens regarding to participating in storage network.
-#[derive(PartialEq, Eq, Clone, Encode, Decode, RuntimeDebug, TypeInfo, Default)]
-pub struct StoragePrefs {
-	foo: bool, // ToDo
-}
-
-/// Preference of what happens regarding to participating in CDN.
-#[derive(PartialEq, Eq, Clone, Encode, Decode, RuntimeDebug, TypeInfo, Default)]
-pub struct EdgePrefs {
-	foo: bool, // ToDo
-}
-
 #[frame_support::pallet]
 pub mod pallet {
 	use super::*;
@@ -142,12 +138,25 @@ pub mod pallet {
 	#[pallet::config]
 	pub trait Config: frame_system::Config {
 		type Currency: LockableCurrency<Self::AccountId, Moment = Self::BlockNumber>;
+
+		/// Default bond size for a CDN participant.
+		#[pallet::constant]
+		type DefaultEdgeBondSize: Get<BalanceOf<Self>>;
+
+		/// Default bond size for a storage network participant.
+		#[pallet::constant]
+		type DefaultStorageBondSize: Get<BalanceOf<Self>>;
+
 		type Event: From<Event<Self>> + IsType<<Self as frame_system::Config>::Event>;
 		/// Number of eras that staked funds must remain bonded for.
 		#[pallet::constant]
 		type BondingDuration: Get<EraIndex>;
 		/// Weight information for extrinsics in this pallet.
 		type WeightInfo: WeightInfo;
+
+		/// Time used for computing era index. It is guaranteed to start being called from the first
+		/// `on_finalize`.
+		type UnixTime: UnixTime;
 	}
 
 	/// Map from all locked "stash" accounts to the controller account.
@@ -155,10 +164,15 @@ pub mod pallet {
 	#[pallet::getter(fn bonded)]
 	pub type Bonded<T: Config> = StorageMap<_, Twox64Concat, T::AccountId, T::AccountId>;
 
-	/// The bond size required to become and maintain the role of a CDN or storage network
-	/// participant.
+	/// The bond size required to become and maintain the role of a CDN participant.
 	#[pallet::storage]
-	pub type BondSize<T: Config> = StorageValue<_, BalanceOf<T>, ValueQuery>;
+	pub type EdgeBondSize<T: Config> =
+		StorageValue<_, BalanceOf<T>, ValueQuery, T::DefaultEdgeBondSize>;
+
+	/// The bond size required to become and maintain the role of a storage network participant.
+	#[pallet::storage]
+	pub type StorageBondSize<T: Config> =
+		StorageValue<_, BalanceOf<T>, ValueQuery, T::DefaultStorageBondSize>;
 
 	/// Map from all (unlocked) "controller" accounts to the info regarding the staking.
 	#[pallet::storage]
@@ -166,18 +180,15 @@ pub mod pallet {
 	pub type Ledger<T: Config> =
 		StorageMap<_, Blake2_128Concat, T::AccountId, StakingLedger<T::AccountId, BalanceOf<T>>>;
 
-	/// The map from (wannabe) storage network participant stash key to the preferences of that
-	/// storage network participant.
+	/// The list of (wannabe) storage network participants stash keys.
 	#[pallet::storage]
 	#[pallet::getter(fn storages)]
-	pub type Storages<T: Config> =
-		CountedStorageMap<_, Twox64Concat, T::AccountId, StoragePrefs, ValueQuery>;
+	pub type Storages<T: Config> = StorageValue<_, Vec<T::AccountId>, ValueQuery>;
 
-	/// The map from (wannabe) CDN participant stash key to the preferences of that CDN participant.
+	/// The list of (wannabe) CDN participants stash keys.
 	#[pallet::storage]
 	#[pallet::getter(fn edges)]
-	pub type Edges<T: Config> =
-		CountedStorageMap<_, Twox64Concat, T::AccountId, EdgePrefs, ValueQuery>;
+	pub type Edges<T: Config> = StorageValue<_, Vec<T::AccountId>, ValueQuery>;
 
 	/// The current era index.
 	///
@@ -228,6 +239,21 @@ pub mod pallet {
 		AlreadyInRole,
 	}
 
+	#[pallet::hooks]
+	impl<T: Config> Hooks<BlockNumberFor<T>> for Pallet<T> {
+		fn on_finalize(_n: BlockNumberFor<T>) {
+			// Check if we have a new era and if so bump the current era index.
+			let now_as_millis = T::UnixTime::now().as_millis();
+			let computed_era: EraIndex =
+				((now_as_millis - DDC_ERA_START_MS) / DDC_ERA_DURATION_MS) as u32; // saturated
+			if Self::current_era() >= Some(computed_era) {
+				return
+			}
+			CurrentEra::<T>::put(computed_era);
+			// ToDo: add `on_initialize` hook to track `on_finalize` weight
+		}
+	}
+
 	#[pallet::call]
 	impl<T: Config> Pallet<T> {
 		/// Take the origin account as a stash and lock up `value` of its balance. `controller` will
@@ -242,6 +268,7 @@ pub mod pallet {
 		pub fn bond(
 			origin: OriginFor<T>,
 			controller: <T::Lookup as StaticLookup>::Source,
+			#[pallet::compact] value: BalanceOf<T>,
 		) -> DispatchResult {
 			let stash = ensure_signed(origin)?;
 
@@ -255,39 +282,29 @@ pub mod pallet {
 				Err(Error::<T>::AlreadyPaired)?
 			}
 
-			let stash_free = T::Currency::free_balance(&stash);
-
 			// Reject a bond which is considered to be _dust_.
-			if stash_free < T::Currency::minimum_balance() {
-				Err(Error::<T>::InsufficientBond)?
-			}
-
-			let bond_size = BondSize::<T>::get();
-
-			// Reject a bond which is lower then required.
-			if stash_free < bond_size {
+			if value < T::Currency::minimum_balance() {
 				Err(Error::<T>::InsufficientBond)?
 			}
 
 			frame_system::Pallet::<T>::inc_consumers(&stash).map_err(|_| Error::<T>::BadState)?;
 
 			// You're auto-bonded forever, here. We might improve this by only bonding when
-			// you actually store/serve and remove once you unbond.
+			// you actually store/serve and remove once you unbond __everything__.
 			<Bonded<T>>::insert(&stash, &controller);
 
-			Self::deposit_event(Event::<T>::Bonded(stash.clone(), bond_size));
-			let item = StakingLedger {
-				stash,
-				total: bond_size,
-				active: bond_size,
-				unlocking: Default::default(),
-			};
+			let stash_balance = T::Currency::free_balance(&stash);
+			let value = value.min(stash_balance);
+			Self::deposit_event(Event::<T>::Bonded(stash.clone(), value));
+			let item =
+				StakingLedger { stash, total: value, active: value, unlocking: Default::default() };
 			Self::update_ledger(&controller, &item);
 			Ok(())
 		}
 
-		/// Schedule a bond of the stash to be unlocked ready for transfer out after the bond
-		/// period ends.
+		/// Schedule a portion of the stash to be unlocked ready for transfer out after the bond
+		/// period ends. If this leaves an amount actively bonded less than
+		/// T::Currency::minimum_balance(), then it is increased to the full amount.
 		///
 		/// The dispatch origin for this call must be _Signed_ by the controller, not the stash.
 		///
@@ -296,10 +313,7 @@ pub mod pallet {
 		///
 		/// No more than a limited number of unlocking chunks (see `MaxUnlockingChunks`)
 		/// can co-exists at the same time. In that case, [`Call::withdraw_unbonded`] need
-		/// to be called first to remove some of the chunks (if possible). This feature is actually
-		/// not required because we unlock the whole bond at once, means it is impossible to have
-		/// more then one unlocking at time. But this is inherited from the `pallet-staking` and we
-		/// may remove in some future version.
+		/// to be called first to remove some of the chunks (if possible).
 		///
 		/// If a user encounters the `InsufficientBond` error when calling this extrinsic,
 		/// they should call `chill` first in order to free up their bonded funds.
@@ -310,48 +324,58 @@ pub mod pallet {
 		#[pallet::weight(T::WeightInfo::unbond())]
 		pub fn unbond(
 			origin: OriginFor<T>,
+			#[pallet::compact] value: BalanceOf<T>,
 		) -> DispatchResult {
 			let controller = ensure_signed(origin)?;
 			let mut ledger = Self::ledger(&controller).ok_or(Error::<T>::NotController)?;
-
-			if ledger.active.is_zero() {
-				// Nothing to unbond.
-				return Ok(())
-			}
-
 			ensure!(
 				ledger.unlocking.len() < MaxUnlockingChunks::get() as usize,
 				Error::<T>::NoMoreChunks,
 			);
 
-			// Make sure that the user maintains enough active bond for their role.
-			// If a user runs into this error, they should chill first.
-			ensure!(!Storages::<T>::contains_key(&ledger.stash), Error::<T>::InsufficientBond);
-			ensure!(!Edges::<T>::contains_key(&ledger.stash), Error::<T>::InsufficientBond);
+			let mut value = value.min(ledger.active);
 
-			let era = Self::current_era().unwrap_or(0) + T::BondingDuration::get();
+			if !value.is_zero() {
+				ledger.active -= value;
 
-			// Unbond actual active stake instead of the current `BondSize` to allow users bond and
-			// unbond the same amount regardless of changes of the `BondSize`.
-			let unbond_value = ledger.active.clone();
-			ledger.active = Zero::zero();
+				// Avoid there being a dust balance left in the staking system.
+				if ledger.active < T::Currency::minimum_balance() {
+					value += ledger.active;
+					ledger.active = Zero::zero();
+				}
 
-			if let Some(mut chunk) = ledger.unlocking.last_mut().filter(|chunk| chunk.era == era) {
-				// To keep the chunk count down, we only keep one chunk per era. Since
-				// `unlocking` is a FiFo queue, if a chunk exists for `era` we know that it will
-				// be the last one.
-				chunk.value = chunk.value.defensive_saturating_add(unbond_value)
-			} else {
-				ledger
-					.unlocking
-					.try_push(UnlockChunk { value: unbond_value, era })
-					.map_err(|_| Error::<T>::NoMoreChunks)?;
-			};
-			// NOTE: ledger must be updated prior to calling `Self::weight_of`.
-			Self::update_ledger(&controller, &ledger);
+				let min_active_bond = if Edges::<T>::get().contains(&ledger.stash) {
+					EdgeBondSize::<T>::get()
+				} else if Storages::<T>::get().contains(&ledger.stash) {
+					StorageBondSize::<T>::get()
+				} else {
+					Zero::zero()
+				};
 
-			Self::deposit_event(Event::<T>::Unbonded(ledger.stash, unbond_value));
+				// Make sure that the user maintains enough active bond for their role.
+				// If a user runs into this error, they should chill first.
+				ensure!(ledger.active >= min_active_bond, Error::<T>::InsufficientBond);
 
+				// Note: in case there is no current era it is fine to bond one era more.
+				let era = Self::current_era().unwrap_or(0) + T::BondingDuration::get();
+				if let Some(mut chunk) =
+					ledger.unlocking.last_mut().filter(|chunk| chunk.era == era)
+				{
+					// To keep the chunk count down, we only keep one chunk per era. Since
+					// `unlocking` is a FiFo queue, if a chunk exists for `era` we know that it will
+					// be the last one.
+					chunk.value = chunk.value.defensive_saturating_add(value)
+				} else {
+					ledger
+						.unlocking
+						.try_push(UnlockChunk { value, era })
+						.map_err(|_| Error::<T>::NoMoreChunks)?;
+				};
+
+				Self::update_ledger(&controller, &ledger);
+
+				Self::deposit_event(Event::<T>::Unbonded(ledger.stash, value));
+			}
 			Ok(())
 		}
 
@@ -399,43 +423,41 @@ pub mod pallet {
 
 		/// Declare the desire to participate in storage network for the origin controller.
 		///
-		/// Effects will be felt at the beginning of the next era.
-		///
-		/// The dispatch origin for this call must be _Signed_ by the controller, not the stash.
+		/// The dispatch origin for this call must be _Signed_ by the controller, not the stash. The
+		/// bond size must be greater than or equal to the `StorageBondSize`.
 		#[pallet::weight(T::WeightInfo::store())]
-		pub fn store(origin: OriginFor<T>, prefs: StoragePrefs) -> DispatchResult {
+		pub fn store(origin: OriginFor<T>) -> DispatchResult {
 			let controller = ensure_signed(origin)?;
 
 			let ledger = Self::ledger(&controller).ok_or(Error::<T>::NotController)?;
 
-			ensure!(ledger.active >= BondSize::<T>::get(), Error::<T>::InsufficientBond);
+			ensure!(ledger.active >= StorageBondSize::<T>::get(), Error::<T>::InsufficientBond);
 			let stash = &ledger.stash;
 
 			// Can't participate in storage network if already participating in CDN.
-			ensure!(!Edges::<T>::contains_key(&stash), Error::<T>::AlreadyInRole);
+			ensure!(!Edges::<T>::get().contains(&stash), Error::<T>::AlreadyInRole);
 
-			Self::do_add_storage(stash, prefs);
+			Self::do_add_storage(stash);
 			Ok(())
 		}
 
 		/// Declare the desire to participate in CDN for the origin controller.
 		///
-		/// Effects will be felt at the beginning of the next era.
-		///
-		/// The dispatch origin for this call must be _Signed_ by the controller, not the stash.
+		/// The dispatch origin for this call must be _Signed_ by the controller, not the stash. The
+		/// bond size must be greater than or equal to the `EdgeBondSize`.
 		#[pallet::weight(T::WeightInfo::serve())]
-		pub fn serve(origin: OriginFor<T>, prefs: EdgePrefs) -> DispatchResult {
+		pub fn serve(origin: OriginFor<T>) -> DispatchResult {
 			let controller = ensure_signed(origin)?;
 
 			let ledger = Self::ledger(&controller).ok_or(Error::<T>::NotController)?;
 
-			ensure!(ledger.active >= BondSize::<T>::get(), Error::<T>::InsufficientBond);
+			ensure!(ledger.active >= EdgeBondSize::<T>::get(), Error::<T>::InsufficientBond);
 			let stash = &ledger.stash;
 
 			// Can't participate in CDN if already participating in storage network.
-			ensure!(!Storages::<T>::contains_key(&stash), Error::<T>::AlreadyInRole);
+			ensure!(!Storages::<T>::get().contains(&stash), Error::<T>::AlreadyInRole);
 
-			Self::do_add_edge(stash, prefs);
+			Self::do_add_edge(stash);
 			Ok(())
 		}
 
@@ -479,7 +501,8 @@ pub mod pallet {
 
 		/// Update the DDC staking configurations .
 		///
-		/// * `bond_size`: The active bond needed to be a Storage or Edge node.
+		/// * `storage_bond_size`: The active bond needed to be a Storage node.
+		/// * `edge_bond_size`: The active bond needed to be an Edge node.
 		///
 		/// RuntimeOrigin must be Root to call this function.
 		///
@@ -487,7 +510,8 @@ pub mod pallet {
 		#[pallet::weight(10_000)]
 		pub fn set_staking_configs(
 			origin: OriginFor<T>,
-			bond_size: ConfigOp<BalanceOf<T>>,
+			storage_bond_size: ConfigOp<BalanceOf<T>>,
+			edge_bond_size: ConfigOp<BalanceOf<T>>,
 		) -> DispatchResult {
 			ensure_root(origin)?;
 
@@ -501,7 +525,8 @@ pub mod pallet {
 				};
 			}
 
-			config_op_exp!(BondSize<T>, bond_size);
+			config_op_exp!(StorageBondSize<T>, storage_bond_size);
+			config_op_exp!(EdgeBondSize<T>, edge_bond_size);
 
 			Ok(())
 		}
@@ -561,11 +586,11 @@ pub mod pallet {
 		/// NOTE: you must ALWAYS use this function to add a storage network participant to the
 		/// system. Any access to `Storages` outside of this function is almost certainly
 		/// wrong.
-		pub fn do_add_storage(who: &T::AccountId, prefs: StoragePrefs) {
-			Storages::<T>::insert(who, prefs);
+		pub fn do_add_storage(who: &T::AccountId) {
+			Storages::<T>::append(who);
 		}
 
-		/// This function will remove a storage network participant from the `Storages` storage map.
+		/// This function will remove a storage network participant from the `Storages` list.
 		///
 		/// Returns true if `who` was removed from `Storages`, otherwise false.
 		///
@@ -573,14 +598,14 @@ pub mod pallet {
 		/// system. Any access to `Storages` outside of this function is almost certainly
 		/// wrong.
 		pub fn do_remove_storage(who: &T::AccountId) -> bool {
-			let outcome = if Storages::<T>::contains_key(who) {
-				Storages::<T>::remove(who);
-				true
-			} else {
+			Storages::<T>::mutate(|storages| {
+				let maybe_position = storages.iter().position(|s| s == who);
+				if let Some(index) = maybe_position {
+					storages.swap_remove(index);
+					return true
+				};
 				false
-			};
-
-			outcome
+			})
 		}
 
 		/// This function will add a CDN participant to the `Edges` storage map.
@@ -590,11 +615,11 @@ pub mod pallet {
 		/// NOTE: you must ALWAYS use this function to add a CDN participant to the system. Any
 		/// access to `Edges` outside of this function is almost certainly
 		/// wrong.
-		pub fn do_add_edge(who: &T::AccountId, prefs: EdgePrefs) {
-			Edges::<T>::insert(who, prefs);
+		pub fn do_add_edge(who: &T::AccountId) {
+			Edges::<T>::append(who);
 		}
 
-		/// This function will remove a CDN participant from the `Edges` storage map.
+		/// This function will remove a CDN participant from the `Edges` list.
 		///
 		/// Returns true if `who` was removed from `Edges`, otherwise false.
 		///
@@ -602,14 +627,14 @@ pub mod pallet {
 		/// system. Any access to `Edges` outside of this function is almost certainly
 		/// wrong.
 		pub fn do_remove_edge(who: &T::AccountId) -> bool {
-			let outcome = if Edges::<T>::contains_key(who) {
-				Storages::<T>::remove(who);
-				true
-			} else {
+			Edges::<T>::mutate(|storages| {
+				let maybe_position = storages.iter().position(|s| s == who);
+				if let Some(index) = maybe_position {
+					storages.swap_remove(index);
+					return true
+				};
 				false
-			};
-
-			outcome
+			})
 		}
 	}
 }