+++ CHANGELOG.md
@@ -13,7 +13,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [4.3.0]
 
-- ...
+### Added
+
+- DDC Staking `CurrentEra` follows DAC era counter
+
+### Changed
+
+- Preferences parameter removed from `pallet-ddc-staking` calls, `value` parameter returned back to `bond` and `unbond` calls
+- Distinct bond size requirements for `Storage` and `Edge` roles with default to 100 CERE
 
 ## [4.3.0]